Full cleanup and format. Remove all warnings.

+++ src/main/java/com/cricketcraft/chisel/block/BlockCarvablePumpkin.java
@@ -36,65 +36,65 @@ public class BlockCarvablePumpkin extends BlockPumpkin implements ICarvable {
 		carverHelper = new CarvableHelper(this);
 	}
 
-    @Override
-    public void onBlockAdded(World world, int x, int y, int z){
-        if(world.getBlock(x, y - 1, z) == Blocks.snow && world.getBlock(x, y - 2, z) == Blocks.snow){
-            if(!world.isRemote){
-                //Let's grab the pumpkin before we start
-                ItemStack pumpkin = new ItemStack(world.getBlock(x, y, z), world.getBlockMetadata(x, y, z));
+	@Override
+	public void onBlockAdded(World world, int x, int y, int z) {
+		if (world.getBlock(x, y - 1, z) == Blocks.snow && world.getBlock(x, y - 2, z) == Blocks.snow) {
+			if (!world.isRemote) {
+				// Let's grab the pumpkin before we start
+				ItemStack pumpkin = new ItemStack(world.getBlock(x, y, z), world.getBlockMetadata(x, y, z));
 
-                world.setBlock(x, y, z, Blocks.air, 0, 2);
-                world.setBlock(x, y - 1, z, Blocks.air, 0, 2);
-                world.setBlock(x, y - 2, z, Blocks.air, 0, 2);
-                EntityChiselSnowman snowman = new EntityChiselSnowman(world);
-                snowman.setCurrentItemOrArmor(2, pumpkin);
-                snowman.setLocationAndAngles((double) x + 0.5D, (double) y - 1.95D, (double) z + 0.5D, 0.0F, 0.0F);
-                world.spawnEntityInWorld(snowman);
-                world.notifyBlockChange(x, y, z, Blocks.air);
-                world.notifyBlockChange(x, y - 1, z, Blocks.air);
-                world.notifyBlockChange(x, y - 2, z, Blocks.air);
-            }
+				world.setBlock(x, y, z, Blocks.air, 0, 2);
+				world.setBlock(x, y - 1, z, Blocks.air, 0, 2);
+				world.setBlock(x, y - 2, z, Blocks.air, 0, 2);
+				EntityChiselSnowman snowman = new EntityChiselSnowman(world);
+				snowman.setCurrentItemOrArmor(2, pumpkin);
+				snowman.setLocationAndAngles(x + 0.5D, y - 1.95D, z + 0.5D, 0.0F, 0.0F);
+				world.spawnEntityInWorld(snowman);
+				world.notifyBlockChange(x, y, z, Blocks.air);
+				world.notifyBlockChange(x, y - 1, z, Blocks.air);
+				world.notifyBlockChange(x, y - 2, z, Blocks.air);
+			}
 
-            //Spawn some lovely particles
-            for(int c = 0; c < 120; ++c){
-                world.spawnParticle("snowshovel", (double)x + world.rand.nextDouble(), (double) (y - 2) + world.rand.nextDouble() * 2.5D, (double) z + world.rand.nextDouble(), 0.0D, 0.0D, 0.0D);
-            }
-        } else if(world.getBlock(x, y - 1, z) == Blocks.iron_block || world.getBlock(x, y - 2, z) == Blocks.iron_block){
-            boolean flag = world.getBlock(x - 1, y - 1, z) == Blocks.iron_block && world.getBlock(x + 1, y - 1, z) == Blocks.iron_block;
-            boolean flag1 = world.getBlock(x, y - 1, z - 1) == Blocks.iron_block && world.getBlock(x, y - 1, z + 1) == Blocks.iron_block;
+			// Spawn some lovely particles
+			for (int c = 0; c < 120; ++c) {
+				world.spawnParticle("snowshovel", x + world.rand.nextDouble(), y - 2 + world.rand.nextDouble() * 2.5D, z + world.rand.nextDouble(), 0.0D, 0.0D, 0.0D);
+			}
+		} else if (world.getBlock(x, y - 1, z) == Blocks.iron_block || world.getBlock(x, y - 2, z) == Blocks.iron_block) {
+			boolean flag = world.getBlock(x - 1, y - 1, z) == Blocks.iron_block && world.getBlock(x + 1, y - 1, z) == Blocks.iron_block;
+			boolean flag1 = world.getBlock(x, y - 1, z - 1) == Blocks.iron_block && world.getBlock(x, y - 1, z + 1) == Blocks.iron_block;
 
-            if(flag || flag1){
-                world.setBlock(x, y, z, Blocks.air, 0, 2);
-                world.setBlock(x, y - 1, z, Blocks.air, 0, 2);
-                world.setBlock(x, y - 2, z, Blocks.air, 0, 2);
+			if (flag || flag1) {
+				world.setBlock(x, y, z, Blocks.air, 0, 2);
+				world.setBlock(x, y - 1, z, Blocks.air, 0, 2);
+				world.setBlock(x, y - 2, z, Blocks.air, 0, 2);
 
-                if(flag){
-                    world.setBlock(x - 1, y - 1, z, Blocks.air, 0, 2);
-                    world.setBlock(x + 1, y - 1, z, Blocks.air, 0, 2);
-                } else {
-                    world.setBlock(x, y - 1, z - 1, Blocks.air, 0, 2);
-                    world.setBlock(x, y - 1, z + 1, Blocks.air, 0, 2);
-                }
+				if (flag) {
+					world.setBlock(x - 1, y - 1, z, Blocks.air, 0, 2);
+					world.setBlock(x + 1, y - 1, z, Blocks.air, 0, 2);
+				} else {
+					world.setBlock(x, y - 1, z - 1, Blocks.air, 0, 2);
+					world.setBlock(x, y - 1, z + 1, Blocks.air, 0, 2);
+				}
 
-                EntityIronGolem ironGolem = new EntityIronGolem(world);
-                ironGolem.setPlayerCreated(true);
-                ironGolem.setLocationAndAngles((double) x + 0.5D, (double) y - 1.95D, (double) z + 0.5D, 0.0F, 0.0F);
-                world.spawnEntityInWorld(ironGolem);
+				EntityIronGolem ironGolem = new EntityIronGolem(world);
+				ironGolem.setPlayerCreated(true);
+				ironGolem.setLocationAndAngles(x + 0.5D, y - 1.95D, z + 0.5D, 0.0F, 0.0F);
+				world.spawnEntityInWorld(ironGolem);
 
-                world.notifyBlockChange(x, y, z, Blocks.air);
-                world.notifyBlockChange(x, y - 1, z, Blocks.air);
-                world.notifyBlockChange(x, y - 2, z, Blocks.air);
+				world.notifyBlockChange(x, y, z, Blocks.air);
+				world.notifyBlockChange(x, y - 1, z, Blocks.air);
+				world.notifyBlockChange(x, y - 2, z, Blocks.air);
 
-                if(flag){
-                    world.notifyBlockChange(x - 1, y - 1, z, Blocks.air);
-                    world.notifyBlockChange(x + 1, y - 1, z, Blocks.air);
-                } else {
-                    world.notifyBlockChange(x, y - 1, z - 1, Blocks.air);
-                    world.notifyBlockChange(x, y - 1, z + 1, Blocks.air);
-                }
-            }
-        }
-    }
+				if (flag) {
+					world.notifyBlockChange(x - 1, y - 1, z, Blocks.air);
+					world.notifyBlockChange(x + 1, y - 1, z, Blocks.air);
+				} else {
+					world.notifyBlockChange(x, y - 1, z - 1, Blocks.air);
+					world.notifyBlockChange(x, y - 1, z + 1, Blocks.air);
+				}
+			}
+		}
+	}
 
 	@Override
 	@SideOnly(Side.CLIENT)
